Add support for managing notes about tasks
Tasks not able to contain additional information.

Tasks should have the ability to contain
additional information.

Let's have each task contain a Note object with
text as information.

Using a Note object instead of a simple String
variable allows for more information to be
attached to the task using the Note object in the
future.